# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Restaurant API",
        "contact": {
            "name": "Hojune Yoo",
            "email": "soy0803@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get all restaurants",
                "operationId": "GetAllRestaurants",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Save a restaurant",
                "operationId": "SaveRestaurant",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Restaurant name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address_line1",
                        "in": "formData",
                        "description": "address line 1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address_line2",
                        "in": "formData",
                        "description": "address line 2",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "city",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "formData",
                        "description": "state",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "zip_code",
                        "in": "formData",
                        "description": "zip code",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "owner's user Id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "API Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/restaurants/{restaurantId}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a Restaurant",
                "operationId": "GetRestaurant",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "description": "Restaurant id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "API Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update a restaurant",
                "operationId": "UpdateRestaurant",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Restaurant name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address_line1",
                        "in": "formData",
                        "description": "address line 1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address_line2",
                        "in": "formData",
                        "description": "address line 2",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "city",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "formData",
                        "description": "state",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "zip_code",
                        "in": "formData",
                        "description": "zip code",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "description": "Restaurant Id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "API Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a Restaurant",
                "operationId": "DeleteRestaurant",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "description": "Restaurant id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "API Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/restaurants/menus/{restaurantId}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a Restaurant with menu",
                "operationId": "GetRestaurantWithMenu",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "description": "Restaurant id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "API Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}